q3/requirements.txt
fastapi[standard]
uvicorn
python-multipart
jinja2
python-dotenv
google-generativeai
pydantic
typing-extensions
aiofiles